Modifications in ETS for issue #117
Some modifications were needed in the ETS code. The validations have been divided in two main blocks of code:

In a first step:
We validate that exists at least one url that is a valid service (WFS, WMS, WCS, SOS or Atom).

In a second step:
If one or more URL exists and is valid, there will be no 'TR.unknownXMLResource' errors.

If there is not any available URL, some warning(s) will be shown.