Modifications in the ETS to solve the issue #117
The validations have been divided in two main blocks of code. In a first step the ETS validates that exists at least one url that is a valid service (WFS, WMS, WCS, SOS or Atom). In a second step, if exist at least one url that fulfils this requirement there will be no 'TR.unknownXMLResource' errors. If there are, the other warning(s) will be shown.