md datasets-and-series 1.5: Spatial Resolution / A double type check is wrong.
An XSD parser valid integer value for double. It also valid scientific notation like 100000e4. INF is also fine. It complains about non numeric.

```
Invalid or missing value. '100000AAA' is not a valid value for (double) in element Distance (gmd:distance).
```

gco:Distance is a double.
http://www.wmo.int/schemas/wmdr/1.0RC7/documentation/schemadoc/schemas/basicTypes_xsd_1/elements/Distance.html

XSD double is defined as https://www.w3.org/TR/xmlschema11-2/#double

So using regex '^-?\d+\.\d{2,}' sounds a bit restrictive to what a double is.

Using cast to double. If failure, then reports
```
[err:FORG0001] Cannot cast to xs:double: DADA.
```

The test is a duplicate with the XSD parsing done in common
requirements.