md datasets-and-series 1.8: Resource Locator / Check that at least one resource locator is defined.
"A Resource locator linking to the service(s) providing online access to the described data set or data
set series shall be given, if such online access is available." does not mean that all gmd:onLine/gmd:URL have to be "valid" URLs.

Report about the expression used to check the URL as it is a bit
restricted vision of what an URL is eg. file://, jdbc://, PRL starting
with // are not considered valid.

When no valid URL provided, then message is now:

```
XML document 'file.xml', record 'c651bc41-d5da-49de-ba47-c51eabbb2007': No online source element found in the record following the rule '(http://|https://|ftp://)([a-z0-9]{1})((\.[a-z0-9-])|([a-z0-9-]))*\.([a-z0-9-]{1,4})(/?)'.
```